\name{analysisHelpers-package}
\alias{analysisHelpers-package}
\alias{analysisHelpers}
\docType{package}
\title{
Simple functions for analyzing behavioral data.
}
\description{
These functions are made to help out with common analysis steps in behavioral experiments.

https://github.com/gbrookshire/analysis_helpers
}
\details{
\tabular{ll}{
Package: \tab analysisHelpers\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-03-21\cr
}

}
\author{
Geoff Brookshire <g_b@cal.berkeley.edu> 
}

%\references{
%~~ Literature or other references for background information ~~
%}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
}
